Revert "feat/240721-friendlist - refactor : refactor the friendList to different files"
This reverts commit 7194c8ec42926bf171e70f3b0f35cf9f89b0086b.
